+++ src/app/app-routing.module.ts
@@ -10,6 +10,10 @@ const routes: Routes = [
     path: 'favorites',
     loadChildren: () => import('@favorite/favorite.module').then(mod => mod.FavoriteModule)
   },
+  {
+    path: 'comics/:id',
+    loadChildren: () => import('@comic/comic.module').then(mod => mod.ComicModule)
+  },
   { path: '', redirectTo: '/home', pathMatch: 'full' },
   { path: '**', redirectTo: '/home', pathMatch: 'full' }
 ];