+++ src/app/feature/comic/comic-routing.module.ts
@@ -1,0 +1,16 @@
+import { NgModule } from '@angular/core';
+import { RouterModule, Routes } from '@angular/router';
+import { ComicComponent } from './comic.component';
+
+const routes: Routes = [
+  {
+    path: '',
+    component: ComicComponent
+  }
+];
+
+@NgModule({
+  imports: [RouterModule.forChild(routes)],
+  exports: [RouterModule]
+})
+export class ComicRoutingModule { }